my server now works
